v1.2

+++ Launcher.cs
@@ -1,19 +1,28 @@
 using Microsoft.Win32;
 using System.ComponentModel;
+using System.Configuration;
 using System.Diagnostics;
 
 namespace RageMPdevLauncher
 {
     public partial class RAGEMPdevLauncher : Form
     {
+        string subKeyPath = "Software\\RAGE-MP";
+        string? legacyPath = ConfigurationManager.AppSettings["legacy_path"];
+        string? enhancedPath = ConfigurationManager.AppSettings["enhanced_path"];
+        string? ragempPath = ConfigurationManager.AppSettings["ragemp_path"];
+        string? cefPort = ConfigurationManager.AppSettings["cef_port"];
         public RAGEMPdevLauncher()
         {
             InitializeComponent();
             string? game_v_path = null;
             string? cefPort = null;
+            ragemp_path_label.Text = ragempPath != null ? ragempPath : "Not Set";
+            legacyPathLabel.Text = legacyPath != null ? legacyPath : "Not Set";
+            enhancedPathLabel.Text = enhancedPath != null ? enhancedPath : "Not Set";
             try
             {
-                using (RegistryKey? key = Registry.CurrentUser.OpenSubKey("Software\\RAGE-MP"))
+                using (RegistryKey? key = Registry.CurrentUser.OpenSubKey(subKeyPath))
                 {
                     if (key != null)
                     {
@@ -25,13 +34,13 @@ namespace RageMPdevLauncher
                         }
                         if (game_v_path != null)
                         {
-                            if (game_v_path.Contains("legacy"))
+                            if (game_v_path.Contains("Enhanced"))
                             {
-                                legacy.Checked = true;
+                                enhanced.Checked = true;
                             }
                             else
                             {
-                                enhanced.Checked = true;
+                                legacy.Checked = true;
                             }
                         }
                         else
@@ -39,6 +48,8 @@ namespace RageMPdevLauncher
                             legacy.Checked = true;
                         }
                     }
+                    else
+                        MessageBox.Show("An error occurred while reading the registry: RageMP key not found");
                 }
             }
             catch (Exception ex)
@@ -49,10 +60,15 @@ namespace RageMPdevLauncher
 
         private void launch_Click(object sender, EventArgs e)
         {
+            if (string.IsNullOrEmpty(ragempPath))
+            {
+                MessageBox.Show("RAGEMP Path is not set in the configuration file.");
+                return;
+            }
             const int ERROR_CANCELLED = 1223; //The operation was canceled by the user.
 
-            ProcessStartInfo info = new ProcessStartInfo(@"C:\RAGEMP\updater.exe");
-            info.WorkingDirectory = @"C:\RAGEMP\";
+            ProcessStartInfo info = new ProcessStartInfo($@"{ragempPath}\updater.exe");
+            info.WorkingDirectory = ragempPath;
             info.UseShellExecute = true;
             info.Verb = "runas";
             try
@@ -72,10 +88,15 @@ namespace RageMPdevLauncher
         {
             try
             {
-                using RegistryKey? myKey = Registry.CurrentUser.OpenSubKey(@"Software\RAGE-MP", true);
+                if (string.IsNullOrEmpty(legacyPath))
+                {
+                    MessageBox.Show("Legacy Path is not set in the configuration file.");
+                    return;
+                }
+                using RegistryKey? myKey = Registry.CurrentUser.OpenSubKey(subKeyPath, true);
                 if (myKey != null)
                 {
-                    myKey.SetValue("game_v_path", "C:\\Program Files (x86)\\Steam\\steamapps\\common\\Grand Theft Auto V", RegistryValueKind.String);
+                    myKey.SetValue("game_v_path", legacyPath, RegistryValueKind.String);
                     myKey.Close();
                 }
             }
@@ -89,10 +110,15 @@ namespace RageMPdevLauncher
         {
             try
             {
-                using RegistryKey? myKey = Registry.CurrentUser.OpenSubKey(@"Software\RAGE-MP", true);
+                if (string.IsNullOrEmpty(enhancedPath))
+                {
+                    MessageBox.Show("Enhanced Path is not set in the configuration file.");
+                    return;
+                }
+                using RegistryKey? myKey = Registry.CurrentUser.OpenSubKey(subKeyPath, true);
                 if (myKey != null)
                 {
-                    myKey.SetValue("game_v_path", "C:\\Program Files (x86)\\Steam\\steamapps\\common\\Grand Theft Auto V Enhanced", RegistryValueKind.String);
+                    myKey.SetValue("game_v_path", enhancedPath, RegistryValueKind.String);
                     myKey.Close();
                 }
             }
@@ -106,18 +132,23 @@ namespace RageMPdevLauncher
         {
             try
             {
+                if (string.IsNullOrEmpty(cefPort))
+                {
+                    MessageBox.Show("CEF Port is not set in the configuration file.");
+                    return;
+                }
                 if (cefDebug.Checked)
                 {
-                    using RegistryKey? myKey = Registry.CurrentUser.OpenSubKey(@"Software\RAGE-MP", true);
+                    using RegistryKey? myKey = Registry.CurrentUser.OpenSubKey(subKeyPath, true);
                     if (myKey != null)
                     {
-                        myKey.SetValue("launch.cefPort", "6969", RegistryValueKind.String);
+                        myKey.SetValue("launch.cefPort", cefPort, RegistryValueKind.String);
                         myKey.Close();
                     }
                 }
                 else
                 {
-                    using RegistryKey? myKey = Registry.CurrentUser.OpenSubKey(@"Software\RAGE-MP", true);
+                    using RegistryKey? myKey = Registry.CurrentUser.OpenSubKey(subKeyPath, true);
                     if (myKey != null)
                     {
                         myKey.DeleteValue("launch.cefPort", false);
@@ -128,6 +159,87 @@ namespace RageMPdevLauncher
             catch (Exception ex)  //just for demonstration...it's always best to handle specific exceptions
             {
                 MessageBox.Show("An error occurred while setting the registry: " + ex.Message);
+            }
+        }
+
+        private void settingsRagePath_Click(object sender, EventArgs e)
+        {
+            FolderBrowserDialog folderDlg = new FolderBrowserDialog();
+            folderDlg.ShowNewFolderButton = false;
+            folderDlg.UseDescriptionForTitle = true;
+            folderDlg.Description = "Select your RAGE MP folder (where updater.exe is located)";
+            folderDlg.SelectedPath = ragempPath != null ? ragempPath : Environment.SpecialFolder.MyComputer.ToString();
+            DialogResult result = folderDlg.ShowDialog();
+            if (result == DialogResult.OK)
+            {
+                if (File.Exists(folderDlg.SelectedPath + "/updater.exe"))
+                {
+                    MessageBox.Show("RAGE MP Path set to: " + folderDlg.SelectedPath);
+                    ragempPath = folderDlg.SelectedPath;
+                    ConfigurationManager.AppSettings["ragemp_path"] = ragempPath;
+                }
+                else
+                {
+                    MessageBox.Show("The selected folder does not contain updater.exe. Please select the correct RAGE MP folder.");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Cancelled, the path was not changed.");
+            }
+        }
+
+        private void settingsEnhancedPath_Click(object sender, EventArgs e)
+        {
+            FolderBrowserDialog folderDlg = new FolderBrowserDialog();
+            folderDlg.ShowNewFolderButton = false;
+            folderDlg.UseDescriptionForTitle = true;
+            folderDlg.Description = "Select your GTA V Enhanced folder (where GTA5_Enhanced.exe is located)";
+            folderDlg.SelectedPath = enhancedPath != null ? enhancedPath : Environment.SpecialFolder.MyComputer.ToString();
+            DialogResult result = folderDlg.ShowDialog();
+            if (result == DialogResult.OK)
+            {
+                if (File.Exists(folderDlg.SelectedPath + "/GTA5_Enhanced.exe"))
+                {
+                    MessageBox.Show("GTA V Enhanced Path set to: " + folderDlg.SelectedPath);
+                    enhancedPath = folderDlg.SelectedPath;
+                    ConfigurationManager.AppSettings["enhanced_path"] = enhancedPath;
+                }
+                else
+                {
+                    MessageBox.Show("The selected folder does not contain GTA5_Enhanced.exe. Please select the correct GTA V Enhanced folder.");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Cancelled, the path was not changed.");
+            }
+        }
+
+        private void settingsLegacyPath_Click(object sender, EventArgs e)
+        {
+            FolderBrowserDialog folderDlg = new FolderBrowserDialog();
+            folderDlg.ShowNewFolderButton = false;
+            folderDlg.UseDescriptionForTitle = true;
+            folderDlg.Description = "Select your GTA V Legacy folder (where GTA5.exe is located)";
+            folderDlg.SelectedPath = legacyPath != null ? legacyPath : Environment.SpecialFolder.MyComputer.ToString();
+            DialogResult result = folderDlg.ShowDialog();
+            if (result == DialogResult.OK)
+            {
+                if (File.Exists(folderDlg.SelectedPath + "/GTA5.exe"))
+                {
+                    MessageBox.Show("GTA V Legacy Path set to: " + folderDlg.SelectedPath);
+                    legacyPath = folderDlg.SelectedPath;
+                    ConfigurationManager.AppSettings["legacy_path"] = legacyPath;
+                }
+                else
+                {
+                    MessageBox.Show("The selected folder does not contain GTA5.exe. Please select the correct GTA V Legacy folder.");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Cancelled, the path was not changed.");
             }
         }
     }
